Fix wallpaper display issue
The Launcher left-aligns built-in wallpapers, which we did not plan
for when designing the new wallpapers. This commit introduces a
slightly modified default wallpaper that looks better when left-aligned.

Now including a thumbnail, so that the cached thumbnail of the previous
default wallpaper won't show up in the wallpaper picker when upgrading
from Android L.

FPIIM-616

Change-Id: I5d14b0471c494f88ff8e6c3d40d9f0f1b954125e