Only wipe cached resolution sizes for swapped camera
Invalidate the cached resolution sizes only for the swapped cameras.
The previous behaviour was to wipe the cached resolution sizes for both
cameras whenever one of them was changed. Considering that this cache
exists to limit the resources spent on accessing the camera capabilities,
do not wipe it if it is not needed.

The "any camera changed" property is now useless, no need to read it.

FPIIM-1291

Change-Id: I0ac53d86b76469f734467fb0a9bb36a3c99f29bf