Re-enable QC AVC decoders
Increase the maximum amount of memory mediaserver is allowed to
allocate and re-enable Qualcomm's AVC decoders.
This commit effectively increases the percentage of total memory that
mediaserver is allowed to allocate from 65 to 75 percent, which seems
to be the sweet spot to get the android.media.cts.ResourceManagerTest
class to pass.

FPIIM-774

Change-Id: I3064560686bcdb1583062193e040c0e02c048d28